Moved plot_sample_qc logic to cpp (cherry picked from commit cbac8f5)
